% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_mat.R
\name{corrupt_mat}
\alias{corrupt_mat}
\title{Corrupts a data matrix by subjecting specified columns to a given limit of detection.}
\usage{
corrupt_mat(mat, cols, limit, fill = "NA")
}
\arguments{
\item{mat}{The data matrix to be corrupted.}

\item{cols}{The columns in the data matrix to subject to an LOD.}

\item{limit}{The percent of each specified column to put under the LOD.
Ex: if \code{limit = 0.25} then the first quartile of each specified column is put under the LOD.}

\item{fill}{The impute scheme to use for values below LOD. Takes one of the following:
\code{"sqrt2", "-1", "mean", "NA", "status"} (Default = \code{"NA"}). See below for descriptions of each.}
}
\value{
The corrupted matrix
}
\description{
\code{corrupt_mat} subjects specified columns of a given data matrix to a given
limit of detection (LOD). Follows a given imputation scheme when imputing values below LOD.
}
\section{Options for \code{fill}}{

\describe{
  \item{\code{"sqrt2"}}{results in values below the LOD imputed as \code{LOD/sqrt\{2\}}.}
  \item{\code{"-1"}}{results in values below the LOD imputed as \code{-1}.}
  \item{\code{"mean"}}{imputes values below the LOD with the mean of those values below the LOD.}
  \item{\code{"NA"}}{replaces values below the LOD with \code{NA} missing values.}
  \item{\code{"status"}}{is a special option that returns a vector detailing whether or not values are below the given LOD or not.}
}
Note that the LOD is column-specific.
}

\examples{
mat <- matrix(rnorm(25), 5, 5)
corrupted_mat <- corrupt_mat(mat, cols=1:5, limit=.4, fill="-1")
}
\seealso{
\code{\link{impute}}
}
