% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrupt_mat_randomly.R
\name{corrupt_mat_randomly}
\alias{corrupt_mat_randomly}
\title{Corrupts a data matrix by knocking out a random specified percentage of entries.}
\usage{
corrupt_mat_randomly(mat, k, perc_b)
}
\arguments{
\item{mat}{The data matrix to be corrupted.}

\item{k}{The random seed to use.}

\item{perc_b}{The percentage of the entries in the data matrix to be corrupted.}
}
\value{
A list containing the corrupted matrix (named \code{cor.mat}), along with a binary matrix (named \code{cor.mask}) where 1's identify entries that were corrupted, and 0
identify entries that were left untouched.
}
\description{
\code{corrupt_mat_randomly} corrupts a given data matrix by knocking out a random
specified percentage of entries as \code{NA} missing values.
}
\examples{
mat <- matrix(rnorm(25), 5, 5)
corrupted_mat <- corrupt_mat_randomly(mat, k = 1, perc_b = .2)
}
\seealso{
\code{\link{corrupt_mat}}
}
