% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimate_mat.R
\name{decimate_mat}
\alias{decimate_mat}
\title{Decimates a data matrix based on the patterns the data matrix was simulated from.}
\usage{
decimate_mat(mat, patterns)
}
\arguments{
\item{mat}{The matrix to be decimated.}

\item{patterns}{The ground truth patterns the matrix was simulated from.}
}
\value{
The decimated matrix
}
\description{
Given a data matrix, \eqn{M}, simulated from a ground-truth pattern matrix, \eqn{P},
\code{decimate_mat} corrupts \eqn{M} row by row to produce \eqn{M}', such that every row of \eqn{M}'
has every value missing besides exactly one random observation corresponding to each pattern in \eqn{P}
(where the patterns are the rows of \eqn{P}).
}
\examples{
pattern <- t(cbind(c(1,1,1,0,0), c(0,0,0,1,1)))
scores <- matrix(rnorm(10), 5, 2)
sim_mat <- scores \%*\% pattern
decimated_mat <- decimate_mat(sim_mat,  patterns=pattern)
}
\seealso{
\code{\link{decimate_row}}
}
