% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimate_row.R
\name{decimate_row}
\alias{decimate_row}
\title{Decimates a given row of a data matrix based on the patterns the data matrix was simulated from.}
\usage{
decimate_row(row, pats)
}
\arguments{
\item{row}{The row to be decimated.}

\item{pats}{The ground truth patterns the row was simulated from.}
}
\value{
The decimated row.
}
\description{
Given a row vector, \eqn{\beta}, simulated from a ground-truth pattern matrix, \eqn{P},
\code{decimate_row} corrupts \eqn{\beta} to produce \eqn{\beta}', such that \eqn{\beta}'
has every value missing besides exactly one random observation corresponding to each pattern in \eqn{P}
(where the patterns are the rows of \eqn{P}).
}
\seealso{
\code{\link{decimate_mat}}
}
\keyword{internal}
