% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{A rough function to perform exploratory data analysis on a given matrix. Returns correlation matrix and PCA summary figures.}
\usage{
eda(
  mat,
  cor_method = c("pairwise.complete.obs", "pearson"),
  cor_lbl = FALSE,
  impute = 0,
  scale_flag = TRUE,
  pcs = c("PC1", "PC2", "PC3", "PC4", "PC5", "PC6"),
  colgroups = NULL,
  rowgroups = NULL,
  rowgroups_name = "cohort",
  frame = F
)
}
\arguments{
\item{mat}{The matrix to perform the eda on.}

\item{cor_method}{Arguments to be passed to \code{GGally::ggcorr}'s "method" argument. By default, \code{cor_method = c("pairwise.complete.obs", "pearson")}.}

\item{cor_lbl}{A logical indicating if there should be labels on the correlation matrix. By default, \code{cor_lbl = FALSE}.}

\item{impute}{What to impute missing values with. By default, \code{impute = 0}.}

\item{scale_flag}{A logical indicating if the matrix should be scaled before running PCA. By default, \code{scale_flag = TRUE}.}

\item{pcs}{A character vector detailing which components you would like to get the loadings plots for. By default, \code{pcs = c("PC1", "PC2", "PC3", "PC4", "PC5", "PC6")}.}

\item{colgroups}{An optional dataframe with the column names for \code{mat} in the first column, and their grouping in the second column. By default, \code{colgroups = NULL}.}

\item{rowgroups}{An optional vector identifying the groups the rows of \code{mat} belong to. Its length should be = \code{nrow(mat)}. By default, \code{rowgroups = NULL}.}

\item{rowgroups_name}{A character providing the overall name for the \code{rowgroups} vector.}

\item{frame}{An optional logical specifying if you would like frames over the clusters in a PCA biplot. By default, \code{frame = FALSE}. Only worth passing as \code{TRUE} when \code{rowgroups} is not \code{NULL}.}
}
\value{
A list containing the following:
\describe{
   \item{\code{cor}}{A correlation matrix figure. From \code{GGally::ggcorr}.}
   \item{\code{var}}{A summary table/figure describing PCA's principle components and their respective proportion of variance explained. From the \code{kableExtra} package.}
   \item{\code{var_raw}}{The raw data underlying the \code{var} figure.}
   \item{\code{load}}{A figure of the loadings for the first \code{pcs}-many components PCA has to offer. With the default value of \code{pcs}, you get the loadings for the first six components.}
   \item{\code{biplot1}}{A PCA biplot figure of PC1 plotted against PC2.}
   \item{\code{biplot2}}{A PCA biplot figure of PC1 plotted against PC3. \code{NULL} if there are not at least 3 PCs returned by PCA.}
   \item{\code{biplot3}}{A PCA biplot figure of PC2 plotted against PC3. \code{NULL} if there are not at least 3 PCs returned by PCA.}
}
}
\description{
\code{eda} is a (very) rough function to perform exploratory data analysis on a given matrix. Returns correlation matrix and PCA summary figures. 
Note: must have \code{ggfortify} package installed and imported before calling this function.
}
