% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcp_defaults.R
\name{get_pcp_defaults}
\alias{get_pcp_defaults}
\title{Retrieves the default Root PCP parameter settings for a given matrix.}
\usage{
get_pcp_defaults(mat)
}
\arguments{
\item{mat}{The matrix for which the default lambda and mu are needed.}
}
\value{
A list containing the default lambda and mu values, labelled as "lambda" and "mu" respectively.
}
\description{
\code{get_pcp_defaults} retrieves the default Root PCP lambda and mu parameter settings for a given matrix.
}
\examples{

# simulate a data matrix:

n <- 50
p <- 10
data <- sim_data(sim_seed = 1, nrow = n, ncol = p, rank = 3, sigma=0, add_sparse = FALSE)
mat <- data$M

# get the default PCP parameters:
default <- get_pcp_defaults(mat)
default.lambda <- default$lambda
default.mu <- default$mu
}
