% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridize.R
\name{hybridize}
\alias{hybridize}
\title{Pre-processing step for PCP_hybrid}
\usage{
hybridize(D, r, limit)
}
\arguments{
\item{D}{The data matrix to hybridize.}

\item{r}{The rank of the data matrix.}

\item{limit}{The percent of each column in \code{D} to put under the LOD.
Ex: if \code{limit = 0.25} then the first quartile of each column in \code{D}
is put under the LOD. Avoid passing \code{limit = 0} (see \emph{Warnings} below).}
}
\value{
List containing: \code{"M_hybrid", "is_safe", "below_lod_mat"}.
See \strong{Methods} below for further details.
\describe{
  \item{\code{"M_hybrid"}}{The hybridized matrix.}
  \item{\code{"is_safe"}}{A logical vector of length \code{nrows(D)}.
  Entries are \code{TRUE} when the corresponding row in \code{D} is a "safe row",
  and \code{FALSE} when the corresponding row in \code{D} is "unsafe".}
  \item{\code{"below_lod_mat"}}{A binary matrix, where \code{1}'s signify the corresponding
  entry in \code{D} was below LOD, and \code{0}'s signify the corresponding entry was above LOD.}
}
}
\description{
\code{hybridize} computes the necessary information for the PCP_hybrid method.
}
\section{Methods}{

The main idea with PCP_hybrid is to make use of information on "safe" vs. "unsafe" rows in a data matrix.
A \emph{safe row} is defined as a row with at least \code{r}-many entries above the LOD.
An \emph{unsafe row} is a row with less than \code{r}-many entries above the LOD.
\strong{When hybridizing a matrix, if an entry is below the LOD in a safe row, it is imputed as \code{-1},
whereas if it is below the LOD in an unsafe row, it is instead imputed as \code{LOD/sqrt\{2\}}.}
}

\section{Warnings}{

Do not pass \code{limit = 0} to \code{hybridize}, as the underlying \code{\link[stats]{quantile}}
function will result in some values put under the lod anyway (since passing 0 to
\code{\link[stats]{quantile}}) results in the minimum value selected as the LOD.
}

\examples{
data <- sim_data(sim_seed = 1, nrow = 10, ncol = 10, rank = 3, sigma=0, add_sparse = FALSE)
mat <- data$M
hybridize(mat, r = 3, limit = 0.25)
}
