% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Subjects a given vector to a given limit of detection.}
\usage{
impute(v, limit, fill = "NA")
}
\arguments{
\item{v}{The vector to impute (usually a column vector in a matrix).}

\item{limit}{The percent of the vector to put under the LOD.
Ex: if \code{limit = 0.25} then the first quartile of the vector is put under the LOD.}

\item{fill}{The impute scheme to use for values below LOD. Takes one of the following:
\code{"sqrt2", "-1", "mean", "NA", "status"} (Default = \code{"NA"}). See below for descriptions of each.}
}
\value{
The imputed vector.
}
\description{
\code{impute} puts a given vector (usually a column vector) under a given limit of detection
(LOD) by imputing values below the LOD with a given impute scheme.
}
\section{Options for \code{fill}}{

\describe{
  \item{\code{"sqrt2"}}{results in values below the LOD imputed as \code{LOD/sqrt\{2\}}.}
  \item{\code{"-1"}}{results in values below the LOD imputed as \code{-1}.}
  \item{\code{"mean"}}{imputes values below the LOD with the mean of those values below the LOD.}
  \item{\code{"NA"}}{replaces values below the LOD with \code{NA} missing values.}
  \item{\code{"status"}}{is a special option that returns a vector detailing whether or not values are below the given LOD or not.}
}
}

\examples{
mat <- matrix(rnorm(25), 5, 5)
mat[,1] <- impute(mat[,1], limit=0.4, fill="-1")
}
\seealso{
\code{\link{corrupt_mat}}
}
\keyword{internal}
