% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_to_df.R
\name{mat_to_df}
\alias{mat_to_df}
\title{Converts matrix to tibble dataframe}
\usage{
mat_to_df(mat, col_names = NA)
}
\arguments{
\item{mat}{The matrix to be converted to a tibble.}

\item{col_names}{vector of desired column names. By default = "chem01,
chem02, chem03, ..."}
}
\value{
The matrix as a tibble object.
}
\description{
\code{mat_to_df} returns the given matrix as a tibble dataframe.
As a tibble, the structure of the matrix can then be easily viewed with \code{ggplot}.
}
\examples{
mat_as_df <- mat_to_df(matrix(rnorm(9), 3, 3))
}
