% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_patterns.R
\name{print_patterns}
\alias{print_patterns}
\title{Prints a matrix of patterns in a PCA-loadings-style plot.}
\usage{
print_patterns(pats, colgroups = NULL, n = 6, pat_type = "pat", title = "")
}
\arguments{
\item{pats}{A matrix of patterns where each pattern is in a different column.}

\item{colgroups}{An optional dataframe with the column names for \code{pats} in the first column, and their grouping in the second column. By default, \code{colgroups = NULL}.}

\item{n}{The number of patterns to print. By default, \code{n = 6}.}

\item{pat_type}{An optional character specifying the prefix for pattern names. By default, \code{pat_type = "pat"}.}

\item{title}{The title for the plot. By default, \code{title = ""}.}
}
\value{
A ggplot object
}
\description{
\code{print_patterns} is a (very) rough function to print a matrix of patterns in a PCA-style plot.
}
