% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsity.R
\name{sparsity}
\alias{sparsity}
\title{Rough way to determine the sparsity of a given matrix.}
\usage{
sparsity(mat, tol = 1e-05)
}
\arguments{
\item{mat}{The matrix whose sparsity is in question.}

\item{tol}{(optional) a numeric that determines if an entry in \code{mat} is effectively 0. 
If the absolute value of an entry is below \code{tol} then it is effectively 0.}
}
\value{
The sparsity of the given matrix.
}
\description{
\code{sparsity} computes the (crude) sparsity of a given matrix.
}
\examples{
sparsity(matrix(0:8, 3, 3))
}
