% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symsub_dist.R
\name{symsub_dist}
\alias{symsub_dist}
\title{Symmetric Subspace Distance}
\usage{
symsub_dist(U, V)
}
\description{
see: http://www.paper.edu.cn/scholar/showpdf/NUT2cN2INTz0YxeQh
(Wang et al., 2006) for more info
}
\details{
Because all solutions may not successfully identify the correct number of patterns, 
symmetric subspace distance compares matrices of different dimensions 
(e.g., true score matrix 1,000 x 4 vs. solution scores 1,000 x 5). 
Symmetric subspace distance defines a distance measure between two linear subspaces, 
where the number of individual vectors is constant but the number of elements they 
contain may differ. The symmetric distance between m-dimensional 
subspace U and n-dimensional subspace V is defined as 
\code{d(U, V) =sqrt{max (m, n)-sum_{i=1}^{m} sum_{j=1}^{n}left(u_{i}^{mathrm{T}} v_{j}right)^{2}}}
If two subspaces largely overlap, they will have a small distance; if they are almost orthogonal, 
they will have a large distance. To use this method to compare simulations with solutions, 
we took orthonormal bases of all included matrices and calculated the ratio between symmetric subspace
distance and \code{sqrt(max(m,n))}. Two subspaces are similar if and only if this ratio is \code{leq frac{1}{2}}.
}
